B 修正flex架构 构建bison测试体系和编写基本框架 初步实现错误抛出到syntax_error等待后续处理

+++ src/token/token.hpp
@@ -76,7 +76,7 @@ public:
     *   @brief 获取token的int类型值
     *   @return token的int类型值
     */
-    int get_int();
+    long long get_int();
     /*
     *   @brief 获取token的double类型值
     *   @return token的double类型值
